+++ src/main/java/zd/zero/waifu/motivator/plugin/onboarding/UpdateNotification.kt
@@ -4,12 +4,9 @@ import com.intellij.notification.*
 import com.intellij.notification.impl.NotificationsManagerImpl
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.popup.Balloon
-import com.intellij.openapi.wm.IdeFrame
-import com.intellij.openapi.wm.WindowManager
 import com.intellij.ui.BalloonLayoutData
-import com.intellij.ui.awt.RelativePoint
 import zd.zero.waifu.motivator.plugin.WaifuMotivator.PLUGIN_NAME
-import java.awt.Point
+import zd.zero.waifu.motivator.plugin.onboarding.BalloonTools.fetchBalloonParameters
 
 val UPDATE_MESSAGE: String = """
       What's New?<br>
@@ -70,10 +67,4 @@ object UpdateNotification {
 
 }
 
-fun fetchBalloonParameters(project: Project): Pair<IdeFrame, RelativePoint> {
-    val ideFrame =
-        WindowManager.getInstance().getIdeFrame(project) ?: WindowManager.getInstance().allProjectFrames.first()
-    val frameBounds = ideFrame.component.bounds
-    val notificationPosition = RelativePoint(ideFrame.component, Point(frameBounds.x + frameBounds.width, 20))
-    return Pair(ideFrame, notificationPosition)
-}
+
